+++ shared/index.ts
@@ -5,7 +5,6 @@
 
 // Utils
 export { DecayEngine, type DecayProfile, type DecayTier, type DecayEntry, type DecayScore } from './utils/decay';
-export { FileObservationStore, type FileObservation, type ObservationStore, type HotFile, type StalledFile, type TodoHotspot } from './utils/fileObservations';
 export { SecretScanner, SECRET_PATTERNS, type SecretSeverity, type SecretPattern, type SecretFinding, type ScanOptions } from './utils/secretScanner';
 
 // API Clients